+++ src/snakemake/workflow.py
@@ -1324,11 +1324,6 @@ class Workflow(WorkflowExecutorInterface):
             logger.debug(f"shared_storage_local_copies: {shared_storage_local_copies}")
             logger.debug(f"remote_exec: {self.remote_exec}")
             dryrun_or_touch = self.dryrun or self.touch
-            if not dryrun_or_touch and (
-                (self.exec_mode == ExecMode.DEFAULT and shared_storage_local_copies)
-                or (self.remote_exec and not shared_storage_local_copies)
-            ):
-                async_run(self.dag.retrieve_storage_inputs())
 
             should_deploy_sources = (
                 SharedFSUsage.SOURCES not in self.storage_settings.shared_fs_usage
@@ -2085,6 +2080,14 @@ class Workflow(WorkflowExecutorInterface):
             return ruleinfo
 
         return decorate
+
+    @property
+    def runtime_paths(self) -> List[Path]:
+        assert self.storage_settings is not None
+        return [
+            self.storage_settings.local_storage_prefix,
+            self.sourcecache.cache_path,
+        ]
 
     def input(self, *paths, **kwpaths):
         def decorate(ruleinfo):